+++ src/App.js
@@ -3,15 +3,15 @@ import { Route, Routes } from 'react-router-dom';
 import { onAuthStateChanged } from 'firebase/auth';
 import { loadStripe } from '@stripe/stripe-js';
 import { Elements } from '@stripe/react-stripe-js';
-import { auth } from '../../config';
-import { useStateValue } from '../../contexts/StateProvider';
-import Header from '../../layouts/Header';
-import Home from '../../pages/Home';
-import Login from '../../pages/Login';
-import Checkout from '../../pages/Checkout';
-import Payment from '../../pages/Payment';
-import Orders from '../../pages/Orders';
-import './style.css';
+import { auth } from './config';
+import { useStateValue } from './contexts/StateProvider';
+import Header from './layouts/Header';
+import Home from './pages/Home';
+import Login from './pages/Login';
+import Checkout from './pages/Checkout';
+import Payment from './pages/Payment';
+import Orders from './pages/Orders';
+import './components/style.css';
 
 const promise = loadStripe(
   'pk_test_51LrmN4CMUJYb034JpqEEBFp1laHuSskqorlDHr1hAudyNGghY7nmzZMRbou9rQA1flPGNi0CzV4xuONFVvQ9sT2W00svwZtrei'
