--- src/layouts/Header/index.js
@@ -1,64 +1,0 @@
-import { Link } from 'react-router-dom';
-import { signOut } from 'firebase/auth';
-import { auth } from '../../config';
-import { useStateValue } from '../../contexts/StateProvider';
-import SearchIcon from '@mui/icons-material/Search';
-import ShoppingBasketIcon from '@mui/icons-material/ShoppingBasket';
-import './style.css';
-
-function Header() {
-  const [{ basket, user }] = useStateValue();
-
-  const handleAuthentication = () => {
-    if (user) {
-      signOut(auth);
-    }
-  };
-
-  return (
-    <div className="header">
-      <Link to="/">
-        <img
-          className="header__logo"
-          src="http://pngimg.com/uploads/amazon/amazon_PNG11.png"
-          alt="logo"
-        />
-      </Link>
-      <div className="header__search">
-        <input type="text" className="header__searchInput" />
-        <SearchIcon className="header__searchIcon" />
-      </div>
-      <div className="header__nav">
-        <Link to={!user && '/login'}>
-          <div onClick={handleAuthentication} className="header__option">
-            <span className="header__optionLineOne">
-              Hello {user?.email ?? 'Guest'}
-            </span>
-            <span className="header__optionLineTwo">
-              {user ? 'Sign out' : 'Sign In'}
-            </span>
-          </div>
-        </Link>
-        <Link to="/orders">
-          <div className="header__option">
-            <span className="header__optionLineOne">Returns</span>
-            <span className="header__optionLineTwo">& Orders</span>
-          </div>
-        </Link>
-        <div className="header__option">
-          <span className="header__optionLineOne">Yourt</span>
-          <span className="header__optionLineTwo">Prime</span>
-        </div>
-        <Link to="/checkout">
-          <div className="header__optionBasket">
-            <ShoppingBasketIcon />
-            <span className="header__optionLineTwo header__basketCount">
-              {basket.length}
-            </span>
-          </div>
-        </Link>
-      </div>
-    </div>
-  );
-}
-export default Header;
