+++ src/pages/Home.js
@@ -1,4 +1,4 @@
-import Product from '../../components/Product';
+import Product from '../components/Product';
 import './style.css';
 
 function Home() {
