--- src/pages/Login/style.css
@@ -1,63 +1,0 @@
-.login {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-  background-color: white;
-}
-
-.login__logo {
-  margin: 20px auto;
-  object-fit: contain;
-  width: 100px;
-}
-
-.login__container {
-  width: 300px;
-  height: fit-content;
-  display: flex;
-  flex-direction: column;
-  border-radius: 5px;
-  border: 1px solid lightgray;
-  padding: 20px;
-}
-
-.login__container > h1 {
-  font-weight: 500;
-  margin-bottom: 20px;
-}
-
-.login__container > form > h5 {
-  margin-bottom: 5px;
-}
-
-.login__container > form > input {
-  height: 30px;
-  margin-bottom: 10px;
-  background-color: white;
-  width: 98%;
-}
-
-.login__container > p {
-  margin-top: 15px;
-  font-size: 12px;
-}
-
-.login__signInButton {
-  background: #f0c14b;
-  margin-top: 10px;
-  width: 100%;
-  height: 30px;
-  border: 1px solid;
-  border-color: #a88734 #9c7e31 #846a29;
-  border-radius: 2px;
-}
-
-.login__registerButton {
-  margin-top: 10px;
-  width: 100%;
-  height: 30px;
-  border: 1px solid;
-  border-color: darkgray;
-  border-radius: 2px;
-}
