--- src/pages/Orders/style.css
@@ -1,7 +1,0 @@
-.orders {
-  padding: 20px 80px;
-}
-
-.orders > h1 {
-  margin: 30px 0;
-}
